Split tests to reflect code changes
